+++ .gitignore
@@ -15,3 +15,4 @@ hs_err_pid*
 .idea/
 
 demo/saves/
+*.log
