+++ java/src/shared/model/game/Dice.java
@@ -1,39 +1,56 @@
 package shared.model.game;
 
-import java.util.concurrent.ThreadLocalRandom;
+import java.util.List;
 
 /**
- * Representation of game dice
+ * Created by corne on 2/17/2016.
+ *
+ * Representation of Game Dice
  */
-public final class Dice {
-    private int min;
-    private int max;
+public class Dice {
+    private List<Die> dice;
 
     /**
-     * Default Constructor - sets dice range to [1,6]
+     * Default Constructor
+     * @param diceCount Number of die to create
      */
-    public Dice() {
-        this.min = 1;
-        this.max = 6;
+    public Dice(int diceCount){
+        assert diceCount > 0;
+
+        for(int i = 0; i < diceCount; i++){
+            dice.add(new Die()); //Ranges will be [1,6] for each die
+        }
     }
 
     /**
-     * Overloaded Constructor - sets dice range to [min,max]
-     * @param min minimum value the dice can return
-     * @param max maximum value the dice can return
+     * Overloaded Constructor - Allows for specifying die range
+     * @param diceCount Number of die to create
+     * @param dieMin Min value for each die
+     * @param dieMax Max value for each die
      */
-    public Dice(final int min, final int max) {
-        assert max > min;
+    public Dice(int diceCount, int dieMin, int dieMax){
+        assert diceCount > 0;
+        assert dieMax > dieMin;
 
-        this.min = min;
-        this.max = max;
+        for(int i = 0; i < diceCount; i++){
+            dice.add(new Die(dieMin, dieMax));
+        }
     }
 
     /**
-     * Simulates rolling one or more dice
-     * @return Integer value between the min and max (inclusive)
+     * Simulates rolling one or more die
+     * @return value of dice roll (all die added up)
      */
-    public Integer roll() {
-        return ThreadLocalRandom.current().nextInt(this.min, this.max + 1);
+    public int roll(){
+        //Initialize roll value
+        int rollVal = 0;
+
+        //Roll each die
+        for (Die die : dice) {
+            rollVal += die.roll();
+        }
+
+        //Return roll value
+        return rollVal;
     }
 }
